Added check for empty

+++ build_scripts/docker/dockeraws-build.sh
@@ -61,8 +61,10 @@ for app in ${O2_APPS[@]} ; do
 #  for x in `docker images | grep "/${app}.*${TAG}\|/${app}.*none" | awk '{print $3}'`; do 
 #    docker rmi -f $x; 
 #  done
-    x=`docker images | grep "/${app}.*${TAG}\g|/${app}.*none" | awk '{print $3}'`
+  x=`docker images | grep "/${app}.*${TAG}\g|/${app}.*none" | awk '{print $3}'`
+  if [ "${x}" != "" ] ; then
     docker rmi -f $x
+  fi
 done
 
 for app in ${O2_APPS[@]} ; do
